win32.c: suppress a warning
* win32/win32.c (rb_w32_write_console): fix the type to suppress a
  warning of difference in signedness.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@50616 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ win32/win32.c
@@ -6818,7 +6818,7 @@ rb_w32_write_console(uintptr_t strarg, int fd)
 	long curlen = constat_parse(handle, s, (next = ptr, &next), &len);
 	reslen += next - ptr;
 	if (curlen > 0) {
-	    long written;
+	    DWORD written;
 	    if (!WriteConsoleW(handle, ptr, curlen, &written, NULL)) {
 		if (GetLastError() == ERROR_CALL_NOT_IMPLEMENTED)
 		    disable = TRUE;
